separando rutas
