+++ routes/index.js
@@ -20,85 +20,6 @@ router.get('/', function(req, res, next) {
   });
 });
 
-router.get('/travels', auth, function(req, res, next) {
-  Travel.find({
-    'user': req.payload._id
-  }, function(err, travels) {
-    if (err) {
-      return next(err);
-    }
-
-    res.json(travels);
-  });
-});
-
-router.post('/travels', auth, function(req, res, next) {
-  var travel = new Travel(req.body);
-  travel.user = req.payload._id;
-  travel.save(function(err, travel) {
-    if (err) {
-      return next(err);
-    }
-
-    res.json(travel);
-  });
-});
-
-var checkPermission = function(req, res, next) {
-  if ((req.travel && req.travel.user != req.payload._id)
-    || (req.destination && req.destination.user != req.payload._id))
-    return res.status(401).json({
-      message: 'Usuario no autorizado'
-    });
-  else {
-    return next();
-  }
-}
-
-router.delete('/travels/:travel', auth, checkPermission, function(req, res, next) {
-  req.travel.remove(function(err) {
-    if (err) throw err;
-    res.json();
-  });
-});
-
-router.delete('/travels/:travel/:destination/dest', auth, checkPermission, function(req, res, next) {
-  req.destination.remove(function(err) {
-    if (err) throw err;
-    res.json();
-  });
-});
-
-router.get('/travels/:travel', auth, checkPermission, function(req, res, next) {
-  req.travel.populate('destinations', function(err, travel) {
-    if (err) {
-      return next(err);
-    }
-
-    res.json(travel);
-  });
-});
-
-router.post('/travels/:travel/destinations', auth, checkPermission, function(req, res, next) {
-  var destination = new Destination(req.body);
-  destination.travel = req.travel;
-  destination.user = req.payload._id;
-  destination.save(function(err, destination) {
-    if (err) {
-      return next(err);
-    }
-
-    req.travel.destinations.push(destination);
-    req.travel.save(function(err, travel) {
-      if (err) {
-        return next(err);
-      }
-
-      res.json(destination);
-    });
-  });
-});
-
 router.post('/register', function(req, res, next) {
   if (!req.body.username || !req.body.password) {
     return res.status(400).json({
@@ -144,104 +65,5 @@ router.post('/login', function(req, res, next) {
     }
   })(req, res, next);
 });
-
-router.get('/destinations/:destination', auth, checkPermission, function(req, res, next) {
-  req.destination.populate(['pointsOfInterest', 'lodging'], function(err, destination) {
-    if (err) {
-      return next(err);
-    }
-
-    res.json(destination);
-  });
-});
-
-router.post('/destinations/:destination/pointsOfInterest', auth, checkPermission, function(req, res, next) {
-  var pointOfInterest = new PointOfInterest(req.body);
-  pointOfInterest.destination = req.destination;
-
-  pointOfInterest.save(function(err, pointOfInterest) {
-    if (err) {
-      return next(err);
-    }
-
-    req.destination.pointsOfInterest.push(pointOfInterest);
-    req.destination.save(function(err, destination) {
-      if (err) {
-        return next(err);
-      }
-
-      res.json(pointOfInterest);
-    });
-  });
-});
-
-router.post('/destinations/:destination/lodging', auth, checkPermission, function(req, res, next) {
-  var pointOfInterest = new PointOfInterest(req.body);
-
-  pointOfInterest.save(function(err, pointOfInterest) {
-    if (err) {
-      return next(err);
-    }
-
-    req.destination.lodging = pointOfInterest;
-    req.destination.save(function(err, destination) {
-      if (err) {
-        return next(err);
-      }
-
-      res.json(pointOfInterest);
-    });
-  });
-});
-
-router.delete('/destinations/:destination/:pointOfInterest', auth, checkPermission, function(req, res, next) {
-  req.pointOfInterest.remove(function(err) {
-    if (err) throw err;
-    res.json();
-  });
-});
-
-// templates
-
-var myRender = function(view) {
-  return function(req, res) {
-    res.render(view, {
-      title: 'Vacacioner permanentes'
-    });
-  }
-}
-
-router.get('/view/home', myRender('home'))
-router.get('/view/travels/all', myRender('travels/all'))
-router.get('/view/travels/detail', myRender('travels/detail'))
-router.get('/view/login', myRender('users/login'))
-router.get('/view/register', myRender('users/register'))
-router.get('/view/destinations/detail', myRender('destinations/detail'))
-
-var setParam = function(req, res, next, id, model, field) {
-  var query = model.findById(id);
-
-  query.exec(function(err, value) {
-    if (err) {
-      return next(err);
-    }
-    if (!value) {
-      return next(new Error('can\'t find ' + field));
-    }
-
-    req[field] = value;
-    return next();
-  });
-}
-
-var params = function(model, field) {
-  return function(req, res, next, id) {
-    setParam(req, res, next, id, model, field);
-  };
-}
-
-router.param('travel', params(Travel, 'travel'));
-router.param('destination', params(Destination, 'destination'));
-router.param('pointOfInterest', params(PointOfInterest, 'pointOfInterest'));
 
 module.exports = router;
